+++ src/services/DFGFactor.jl
@@ -88,9 +88,9 @@ macro defObservationType(structname, factortype, manifold)
     return esc(
         quote
             # user manifold must be a <:Manifold
-            @assert ($manifold isa AbstractManifold) "@defObservationType manifold (" *
-                                                     string($manifold) *
-                                                     ") is not an `AbstractManifold`"
+            manifoldCheck = ($manifold isa AbstractManifold) || ($manifold isa Function)
+            @assert manifoldCheck string($manifold) *
+                                  " must be an `AbstractManifold` or a `Function` returning one."
 
             @assert ($factortype <: AbstractObservation) "@defObservationType factortype (" *
                                                          string($factortype) *
@@ -100,7 +100,11 @@ macro defObservationType(structname, factortype, manifold)
                 Z::T & (lower = DFG.Packed, choosetype = DFG.resolvePackedType)
             end
 
-            DFG.getManifold(::Type{<:$structname}) = $manifold
+            if $manifold isa AbstractManifold
+                DFG.getManifold(::Type{<:$structname}) = $manifold
+            elseif $manifold isa Function
+                DFG.getManifold(obs::$structname) = $manifold(obs)
+            end
         end,
     )
 end